project-argus-backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
boto3==1.34.34
python-multipart==0.0.6
pydantic==2.5.3
python-dotenv==1.0.0
beautifulsoup4==4.12.3
requests==2.31.0
